Refactor: 코드 리뷰 반영
- types 분리
- filteredCurrentMenu -> findCurrentMenu 네이밍 및 메서드 변경
- recoil export 방식 변경
