+++ src/components/ToolBox/hooks/useToolBox.ts
@@ -1,14 +1,9 @@
 import { useCallback } from "react";
-import { useRecoilState } from "recoil";
-import toolBoxMenus, { ToolBoxMenus } from "@/recoil/toolbox";
-
-interface UseToolBoxReturn {
-  menus: ToolBoxMenus[];
-  handleChangeMenu: (id: number) => void;
-}
+import { UseToolBoxReturn } from "../types";
+import { useToolBoxState } from "@/recoil/toolbox";
 
 const useToolBox = (): UseToolBoxReturn => {
-  const [menus, setMenus] = useRecoilState(toolBoxMenus);
+  const [menus, setMenus] = useToolBoxState();
   const handleChangeMenu = useCallback((id: number) => {
     setMenus(
       menus.map((menu) => {
