+++ src/components/ToolBox/types.ts
@@ -1,0 +1,6 @@
+import { ToolBoxMenus } from "@/recoil/toolbox";
+
+export interface UseToolBoxReturn {
+  menus: ToolBoxMenus[];
+  handleChangeMenu: (id: number) => void;
+}
