+++ src/recoil/toolbox/findCurrentMenu.ts
@@ -1,8 +1,8 @@
 import { selector } from "recoil";
 import toolBoxMenus from "@/recoil/toolbox/atom";
 
-export const filteredCurrentMenu = selector({
-  key: "filteredCurrentMenu",
+export const findCurrentMenu = selector({
+  key: "findCurrentMenu",
   get: ({ get }) => {
     const menus = get(toolBoxMenus);
     return menus.filter((menu) => menu.isActive);
