+++ src/recoil/toolbox/index.ts
@@ -1,5 +1,13 @@
-import atom from "./atom";
-export * from "./atom";
-export { filteredCurrentMenu } from "./filteredCurrentMenu";
+import { useRecoilState, useRecoilValue } from "recoil";
+import toolBoxMenus from "./atom";
+import { findCurrentMenu } from "./findCurrentMenu";
 
-export default atom;
+export * from "./types";
+
+export const useToolBoxState = () => {
+  return useRecoilState(toolBoxMenus);
+};
+
+export const useCurrentMenuValue = () => {
+  return useRecoilValue(findCurrentMenu);
+};
