+++ src/recoil/toolbox/types.ts
@@ -1,0 +1,5 @@
+export interface ToolBoxMenus {
+  id: number;
+  name: string;
+  isActive: boolean;
+}
